move tool_abstract and parse_config into freva_rest/tool_api, merge input parameters into tool_abstract, add title to input params
